move getter, remove todo
